+++ src/Core/Dfe.Complete.Domain/Enums/NoteTaskIdentifier.cs
@@ -27,5 +27,9 @@ public enum NoteTaskIdentifier
 
     [Description("article_of_association")]
     [DisplayDescription("Articles of association")]
-    ArticleOfAssociation = 6
+    ArticleOfAssociation = 6,
+
+    [Description("deed_of_novation_and_variation")]
+    [DisplayDescription("Deed of novation and variation\r\n")]
+    DeedOfNovationAndVariation = 7
 }